pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fortune"
version = "0.1.0"
description = "Secure-by-default runtime scaffold for jobs and services."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "psutil>=7.0.0",
  "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.11.0",
  "pip-audit>=2.8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --cov=fortune --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "C4"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
